Delete fmt::format&fmt::print for C++17
This is an attempt to solve the problem of merging lcenc under ffmpeg in gcc 11.5.0 C⁺⁺17 for amateurs. For GNU 15.0.0 20241020 there is already std::format and std::prrint for me I don't recommend. You have to be able to convert and know what new -f`XXX` functions to add for GNU 15.0.0 to work properly.